test: 호출 제한 통합 테스트 도우미 분리
